+++ src/ui/AudioAnalysisDialogs.h
@@ -4,16 +4,35 @@
 
 #include <QDialog>
 
+#include <functional>
+
 class Database;
+class QFormLayout;
 class QLabel;
 class QPushButton;
 class QTableWidget;
+class QTimer;
 
 class AudioAnalysisStatusDialog final : public QDialog {
     Q_OBJECT
 
 public:
-    explicit AudioAnalysisStatusDialog(const QString &featuresPath, QWidget *parent = nullptr);
+    explicit AudioAnalysisStatusDialog(
+        const QString &featuresPath,
+        std::function<AudioAnalysisData::LiveStatus()> liveStatus = {},
+        QWidget *parent = nullptr);
+
+private:
+    void refreshLiveStatus();
+
+    std::function<AudioAnalysisData::LiveStatus()> m_liveStatus;
+    QLabel *m_liveRunning = nullptr;
+    QLabel *m_liveProgress = nullptr;
+    QLabel *m_liveRate = nullptr;
+    QLabel *m_liveEta = nullptr;
+    QLabel *m_liveElapsed = nullptr;
+    QLabel *m_livePower = nullptr;
+    QTimer *m_liveTimer = nullptr;
 };
 
 class DuplicateCopiesDialog final : public QDialog {